Add Origins of Form
All about form and structure — how things are made, and how they're constrained, from the structural efficiencies and limitations of natural materials, to differences in building/manufacturing practices we might consider in light of what we want to optimize for… It's not only well-written, but graced with lovely illustrations on almost every page, of things ranging from cell structures to ancient tools to the various ways things can break. It also works up to some interesting conclusions…moving from elements of form and structure, to historical and evolutionary influences, the "dictates of chance" on what things exist, and unifying principles of the physical world.