Added a new Book ' Karma Yoga' by Swami Vivekanand
sometimes a thought comes in mind to give up everything, there is so much attachment and suffering in this world. so should one leave everything a become a Sadhu? one should not move away from his responsibilities and one doesn't have to become the sadhu to get the peace. one can find peace in most noisy place in universe and in utmost silence. the book is somewhat different from normal self- help books.  it tells to not be attached to the work  which you are doing. 
it nicely explains the concept of religion and karma, work, why one needs to do the karma.
I feel this is the best book i have ever read.